fetch Explore page data 5 seconds after site load
